Revert "chore: limit usage of gas estimator and connect wallet (#1232)"
This reverts commit 362dac5d49264e532983c60cf1c2d963768aa3df.
