+++ apps/web/app/(basenames)/layout.tsx
@@ -1,4 +1,3 @@
-import CryptoProviders from 'apps/web/app/CryptoProviders';
 import ErrorsProvider from 'apps/web/contexts/Errors';
 import UsernameNav from 'apps/web/src/components/Layout/UsernameNav';
 
@@ -28,12 +27,10 @@ export default async function BasenameLayout({
 }) {
   return (
     <ErrorsProvider context="basenames">
-      <CryptoProviders>
-        <div className="max-w-screen flex min-h-screen flex-col">
-          <UsernameNav />
-          {children}
-        </div>
-      </CryptoProviders>
+      <div className="max-w-screen flex min-h-screen flex-col">
+        <UsernameNav />
+        {children}
+      </div>
     </ErrorsProvider>
   );
 }
