--- apps/web/app/CryptoProviders.tsx
@@ -1,74 +1,0 @@
-'use client';
-
-import '@rainbow-me/rainbowkit/styles.css';
-import '@coinbase/onchainkit/styles.css';
-
-import { OnchainKitProvider } from '@coinbase/onchainkit';
-import { connectorsForWallets, RainbowKitProvider } from '@rainbow-me/rainbowkit';
-import { QueryClient, QueryClientProvider } from '@tanstack/react-query';
-import { isDevelopment } from 'apps/web/src/constants';
-import { createConfig, http, WagmiProvider } from 'wagmi';
-import { base, baseSepolia, mainnet } from 'wagmi/chains';
-import {
-  coinbaseWallet,
-  metaMaskWallet,
-  phantomWallet,
-  rainbowWallet,
-  uniswapWallet,
-  walletConnectWallet,
-} from '@rainbow-me/rainbowkit/wallets';
-
-const connectors = connectorsForWallets(
-  [
-    {
-      groupName: 'Recommended',
-      wallets: [
-        coinbaseWallet,
-        metaMaskWallet,
-        uniswapWallet,
-        rainbowWallet,
-        phantomWallet,
-        walletConnectWallet,
-      ],
-    },
-  ],
-  {
-    projectId: process.env.NEXT_PUBLIC_WALLET_CONNECT_PROJECT_ID ?? 'dummy-id',
-    walletConnectParameters: {},
-    appName: 'Base.org',
-    appDescription: '',
-    appUrl: 'https://www.base.org/',
-    appIcon: '',
-  },
-);
-
-const config = createConfig({
-  connectors,
-  chains: [base, baseSepolia, mainnet],
-  transports: {
-    [base.id]: http(),
-    [baseSepolia.id]: http(),
-    [mainnet.id]: http(),
-  },
-  ssr: true,
-});
-const queryClient = new QueryClient();
-
-type CryptoProvidersProps = {
-  children: React.ReactNode;
-};
-
-export default function CryptoProviders({ children }: CryptoProvidersProps) {
-  return (
-    <WagmiProvider config={config}>
-      <QueryClientProvider client={queryClient}>
-        <OnchainKitProvider
-          chain={isDevelopment ? baseSepolia : base}
-          apiKey={process.env.NEXT_PUBLIC_ONCHAINKIT_API_KEY}
-        >
-          <RainbowKitProvider modalSize="compact">{children}</RainbowKitProvider>
-        </OnchainKitProvider>
-      </QueryClientProvider>
-    </WagmiProvider>
-  );
-}
